Move `useStubsForActions` function to support/helpers
