+++ client-v2/src/app/components/organisms/task/task.component.test.ts
@@ -1,13 +1,12 @@
 import { CdkMenuModule } from '@angular/cdk/menu'
 import { Store } from '@ngrx/store'
-import { testName } from 'cypress/support/helpers'
+import { testName, useStubsForActions } from 'cypress/support/helpers'
 import { FocusableDirective } from 'src/app/directives/focusable.directive'
 import { MutationDirective } from 'src/app/directives/mutation.directive'
 import { EntityType } from 'src/app/fullstack-shared-models/entities.model'
 import { TaskPreviewFlattend, TaskPriority, TaskStatus } from 'src/app/fullstack-shared-models/task.model'
 import { getEntityMenuItemsMap } from 'src/app/shared/entity-menu-items'
 import { AppState } from 'src/app/store'
-import { useStubsForActions } from 'src/app/utils/menu-item.helpers'
 import { IconsModule } from '../../atoms/icons/icons.module'
 import { InlineEditorComponent } from '../../atoms/inline-editor/inline-editor.component'
 import { DropDownComponent } from '../../molecules/drop-down/drop-down.component'
